feat: function for downloading custom url
In the node manager we want to support downloading a 'custom' binary, which is really a binary of
`safenode` from someone's fork. These get uploaded and used during testnet deployments. This new
function downloads the URL not on the basis of a convention, but instead just allowing any URL to be
specified.

We make the assertion that the URL has to point to a zip or gzipped tar file.