feat: support node manager releases
The main noteworthy thing here is that the node manager is in its own repository. This crate was
originally setup to pull releases from the `safe_network` workspace repository, so at the moment the
node manager is a bit of a special case. However, it could be the case that in the future we will
pull binaries from more repositories. This change supports that.

The node manager is a single-crate repository, which makes it much easier to get the latest version:
you can use the specific API for that, rather than processing the whole release list.