refactor: use version type rather than strings
BREAKING CHANGE: the `semver::Version` type is used rather strings for working with versions.

The need for the version parsing test and tests for retrieving the latest version was removed. The
latter of those was only testing that returned string was a valid semantic version, which wouldn't
be possible with a `Version` type.