feat: add node launchpad to release type
A new `node-launchpad` binary is being introduced, which is a node management TUI. It therefore
needs a corresponding `ReleaseType` variant.

The crate for `node-launchpad`, `sn_node_launchpad`, does not yet have a published version, so the
`get_latest_version` function has been hard coded to return the first alpha version. This will be
removed soon.